[bugfix] MatchEndpoint 中新增channel index

+++ src/hixl/cs/channel.h
@@ -30,7 +30,7 @@ struct ChannelDesc {
   uint8_t tc;
   uint8_t sl;
   ChannelType channel_type{ChannelType::kClient};
-  uint32_t channel_index{0U};
+  uint64_t channel_index{0UL};
   uint8_t qos{0U};
 };
